+++ GameTest/Enemy.cpp
@@ -20,14 +20,14 @@ void Enemy::Update(float dt)
 {
 	if (m_sprite)
 	{
-		m_angle = PI - atan2(m_target_y - (double)m_y, (double)m_x - m_target_x);
+		m_angle = PI - atan2f(m_target_y - m_y, m_x - m_target_x);
 		m_sprite->SetAngle(3 * PI / 2 + m_angle);
 		m_sprite->SetPosition(m_x, m_y);
 		m_sprite->Update(dt);
 	}
 	if (m_move_mode == 0)
 	{
-		float d = sqrtf(pow((double)m_target_x - m_x, 2) + pow((double)m_target_y - m_y, 2));
+		float d = sqrtf(powf(m_target_x - m_x, 2) + powf(m_target_y - m_y, 2));
 		float d_p = d / m_initial_d;
 		m_sprite->SetScale(1.0f - d_p);
 		if (d_p > 0.02f)
